+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Inform customers to come 1 hour early to enjoy Aquatonic
+
 ### 1.2.0: January 14th, 2018
 * Force value of downloads_remaining in files
 * Avoid PHP notices and warnings